// src/types/enums/ClientSocketEvents.ts
export enum ClientSocketEvents {
	SetUuid = "setuuid",
	JoinRoom = "joinroom",
	Disconnect = "disconnect",
	Error = "error",
	SyncAllClients = "syncallclients",
	AddClient = "addclient",
	RemoveClient = "removeclient",
	SetPositionOf = "setpositionof",
	SetVentOf = "setventof",
	SetColorOf = "setcolorof",
	SetHost = "sethost",
	SetOptions = "setoptions",
	SetSettings = "setsettings",
	SetGameState = "setgamestate",
	SetGameFlags = "setgameflags",
	SetFlagsOf = "setflagsof",
}